EntityBloodSkeleton Migration + some docs
